Split off the internal turn resolving functions into engine/resolver.py
